Added @RestController annotation in ItemsController. Renamed restcontroller package and classes. Added @JsonManagedReference and @JsonBackReference annotations in models to make them serializable.
